Add files via upload
My assignment was: 

"Make an HTML page with the following two forms:
Two text inputs and one number input. The text inputs will capture a person's first and last name. The number input will capture their age.
Two number inputs. Between the inputs put a plus sign (+) and after the second input put an equal sign (=).
Each of these should be a separate form tag with a separate id. Each of these forms should have a submit button. Add labels to the inputs as appropriate.

Now add a script that performs the following:
When the first form is submitted. Take the data to form a sentence. "My name is [first_name] [last_name] and I am [age] years old." The data in brackets should be replaced with the data from the form. Display this somewhere on the page.
When the second form is submitted, add the two numbers together and show the data on the screen after the equal sign in the form."